+++ taobao88/src/main/java/org/taobao88/taobao/enterprise/dao/BalanceDAO.java
@@ -5,7 +5,7 @@ import org.taobao88.taobao.enterprise.entity.UserT;
 
 public interface BalanceDAO {
 
-	public double getBalance(UserT userT);
+	public int getBalance(UserT userT);
 	
 	public int adjustBalance(BalanceOperation balanceOperation);
