+++ taobao88/src/main/java/org/taobao88/taobao/enterprise/service/BalanceService.java
@@ -5,7 +5,7 @@ import org.taobao88.taobao.enterprise.entity.UserT;
 
 public interface BalanceService {
 
-public double getBalance(UserT userT);
+	public int getBalance(UserT userT);
 	
 	public int adjustBalance(BalanceOperation balanceOperation);
